requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.3 